% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByCov.R
\name{filterByCov}
\alias{filterByCov}
\title{Filter mutation by coverage}
\usage{
filterByCov(VAFdata,VAFcov)
}
\arguments{
\item{VAFdata}{data frame, VAF}

\item{VAFcov}{data frame, VAF cov}
}
\value{
write to output directory
}
\description{
This function filter mutation by coverage
}
\references{
{TBA}
}
