% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMultiSource.R
\name{filterMultiSources}
\alias{filterMultiSources}
\title{Filter multisource contamination}
\usage{
filterMultiSources(filename,output_path,VAFdata,VAFcov,uniq_both)
}
\arguments{
\item{filename}{character, contamination file name}

\item{output_path}{character, output directory}

\item{VAFdata}{data frame, mutation VAF}

\item{VAFcov}{data frame, mutation coverage}

\item{uniq_both}{integer, flag to control type of overlapping mutation in source and target samples}
}
\value{
write contPredict.txt to output_path
}
\description{
This function filter multisource contamination
}
\references{
{TBA}
}
