% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputData.R
\name{inputData}
\alias{inputData}
\title{Read in data from a file}
\usage{
inputData(filename)
}
\arguments{
\item{filename}{character, file name and its path}
}
\value{
data frame
}
\description{
This function reads in file with header
}
\references{
{TBA}
}
