% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_scalar_character}
\alias{is_scalar_character}
\title{Check for scalar argument
This function checks for user's input is a scalar}
\usage{
is_scalar_character(x)
}
\arguments{
\item{x}{character (1)}
}
\value{
boolean TRUE or FALSE
}
\description{
Check for scalar argument
This function checks for user's input is a scalar
}
\details{
user's entry must meet the following criteria:
    character,
    length 1,
    not NA,
    non-zero length
}
