% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleSource.R
\name{multipleSource}
\alias{multipleSource}
\title{Identify true source from multisource contamination}
\usage{
multipleSource(sample_pairs,VAFdata,VAFcov,VAF_cutoff,VAF_cutoff1,p.val_cutoff,output_path)
}
\arguments{
\item{sample_pairs}{data frame, sample infomation}

\item{VAFdata}{data frame, mutation VAF}

\item{VAFcov}{data frame, mutation coverage}

\item{VAF_cutoff}{numeric, minimum VAF (default: 0.002)}

\item{VAF_cutoff1}{numeric, minimum VAF to be considered a source mutation (default: 0.05)}

\item{p.val_cutoff}{numeric, pval cutoff to be considered a significant contamination source (default: 0.05)}

\item{output_path}{character, output directory}
}
\value{
list, remaining paiwise sample relationship information after Fisher exact test that eliminates insignificant contamination pair
}
\description{
This function uses Fisher exact test to filter multiple sources that contaminate a particular target sample
}
\references{
{TBA}
}
