% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numMutationperSample.R
\name{numMutationperSample}
\alias{numMutationperSample}
\title{Count number of SNPs in each sample}
\usage{
numMutationperSample (VAFdata,VAF_cutoff,n)
}
\arguments{
\item{VAFdata}{data frame, variant allele frequency (VAF)}

\item{VAF_cutoff}{numeric, minimum VAF}

\item{n}{numeric, number of sample}
}
\value{
list, SNP count per sample
}
\description{
This function counts number of SNPs in each sample
}
\references{
{TBA}
}
