% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairCountPoint.R
\name{pairCountPoint}
\alias{pairCountPoint}
\title{Count number of SNPs in 7 regions of a given pair of sample}
\usage{
pairCountPoint(VAFdata,pid1,pid2,slope1,slope2)
}
\arguments{
\item{VAFdata}{data frame, Variant Allele Frequency}

\item{pid1}{character, first sample id}

\item{pid2}{character, second sample id}

\item{slope1}{numeric, first slope that is closer to axis-X (default:0.5)}

\item{slope2}{numeric, second slope that is closer to axis-Y (default:2)}
}
\value{
matrix of 7 columns: number of SNPs in 7 regions
}
\description{
This function counts number of SNPs in 7 regions of a given pair of sample
}
\references{
{TBA}
}
