% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairCommon.R
\name{pairPCommon}
\alias{pairPCommon}
\title{Pairwise Sample Function}
\usage{
pairPCommon(VAFdata,VAF_cutoff,num_round_digit,n)
pairShare(VAFdata,VAF_cutoff,n)
pairList(VAFdata,n,del)
pairCommList(SNPcomm,n)
}
\arguments{
\item{VAFdata}{data frame, mutation VAF}

\item{VAF_cutoff}{numeric, minimum VAF}

\item{num_round_digit}{numeric, number of rounding decimal digit (default: 3)}

\item{n}{numeric, number of sample}

\item{SNPcomm}{matrix, output from the calling of pairPCommon(VAFdata,VAF_cutoff,num_round_digit,n)}

\item{del}{character, file delimiter}
}
\value{
pairPCommon: matrix, nxn pcommon, n: number of sample

pairShare: matrix, nxn number of common snps

pairList: vector, all pairwise samples

pairCommList: list, all pairwise samples and their pcommon values
}
\description{
4 different functions for pairwise sample
}
\details{
pairPCommon: calculates pcommon

pairShare: counts number of common SNPs

pairList: generates all pairwise sample list

pairCommList: tabulates pcommon for all pairwise samples
}
\references{
{TBA}
}
