% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos_link.R
\name{plot_circos_link}
\alias{plot_circos_link}
\title{Plot circos with links to show contaminated pairwise sample}
\usage{
plot_circos_link (file,R=300,W=30,plotsize=800,titleStr="",seg.lab.size=1,fig.file="circos.png", contaminatedOnly=TRUE,sameSubjectOnly=FALSE,allSamples=sampleID)
}
\arguments{
\item{file}{character, circos data file name with its path}

\item{R}{integer, circle radius}

\item{W}{integer, segment width}

\item{plotsize}{integer, figure size}

\item{titleStr}{character, title of circos plot}

\item{seg.lab.size}{numeric, segment label font size}

\item{fig.file}{character, circos figure file name}

\item{sameSubjectOnly}{logical, only plot same subject pairs? (default: FALSE)}

\item{allSamples}{character vector, sample ids}

\item{output_path}{character, directory}

\item{contaminationOnly}{logical, only plot contaminated pairs? (default: TRUE)}
}
\value{
.tif figure at output directory
}
\description{
This function reads in circos plotdata and creates circos plot
}
\references{
{TBA}
}
