% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_link.R
\name{circos_link}
\alias{circos_link}
\title{A modified circos plot from OmicCircos package}
\usage{
circos_link (mapping, xc, yc, R, W,cir, type, print.chr.lab, cex,lwd, col, side,scale,seg.lab.size, link.wd,lower_arch,line_type)
}
\arguments{
\item{mapping, }{data frame}

\item{xc}{integer, circle center x coordinate}

\item{yc}{integer, circle center y coordinate}

\item{R}{interger, circle radius}

\item{W}{interger, circle width}

\item{cir, }{data frame}

\item{type}{character, plot type [chr|link]}

\item{print.chr.lab}{logical, TRUE: print chr label, FALSE: not print chr label}

\item{cex}{numeric, plot font size}

\item{lwd}{interger, line width}

\item{col}{vector, colors}

\item{side}{character, segment label position [in|out]}

\item{scale}{logic, scale the circos size (default: FALSE)}

\item{seg.lab.size}{numeric, segment label font size (default:1)}

\item{link.wd}{numeric, link width (default:1)}

\item{lower_arch}{numeric, (default:1)}

\item{line_type}{character, line type (default: "l")}
}
\value{
.tif figure at output directory
}
\description{
To draw circular plot with links shows contamination paired samples
}
\references{
{TBA}
}
