% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByCov.R
\name{filterByCov}
\alias{filterByCov}
\title{Filter mutations by coverage}
\usage{
filterByCov(VAFdata,VAFcov,minCov)
}
\arguments{
\item{VAFdata}{data frame, VAF}

\item{VAFcov}{data frame, VAF cov}

\item{minCov}{numeric, minimum coverage (default:50)}
}
\value{
data frame
}
\description{
This function filters mutation by coverage
}
\references{
{TBA}
}
