% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByCov.R
\name{filterByCov}
\alias{filterByCov}
\title{Filter mutations base on coverage}
\usage{
filterByCov(VAFdata,VAFcov,minCov)
}
\arguments{
\item{VAFdata}{data frame, mutation VAF}

\item{VAFcov}{data frame, mutation coverage}

\item{minCov}{numeric, minimum coverage (default: 50)}
}
\value{
data frame, mutation VAF with coverage > minCov
}
\description{
This function filters mutation base on coverage
}
\references{
{TBA}
}
