% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCommonSNPs.R
\name{filterCommonSNPs}
\alias{filterCommonSNPs}
\title{Filter SNPs base on coefficient of variant}
\usage{
filterCommonSNPs(VAFdata, percentage)
}
\arguments{
\item{VAFdata}{data frame, variant allele frequency (VAF)}

\item{percentage}{numeric, COV cutoff in percentage (default:5)}
}
\value{
integer vector, row index of VAF data to be used
}
\description{
This function filters common SNPs base on coefficient of variant (COV) at the specified cutoff
}
\references{
{TBA}
}
