% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixingRatio.R
\name{mixingRatio}
\alias{mixingRatio}
\title{Calculate contamination level}
\usage{
mixingRatio(VAFdata,VAFcov,sample_pairs,final_rel,VAF_cutoff,VAF_ignore,
ALL_flag,output_path,sameSubject)
}
\arguments{
\item{VAFdata}{data frame, mutation VAF}

\item{VAFcov}{data frame, mutation coverage}

\item{sample_pairs}{data frame, pairwise samples information}

\item{final_rel}{data frame, relation information}

\item{VAF_cutoff}{numeric, minimum VAF (default: 0.002)}

\item{VAF_ignore}{numeric, ignore variant less than this cutoff (default: 0.2)}

\item{ALL_flag}{logical, TRUE: calculate mixing ratio for all pairwise samples, FALSE: calculate mixing ratio for contamination pairwise samples}

\item{output_path}{character, output directory}

\item{sameSubject}{logical, TRUE: calculate contamination level for same subject pairs (default: FALSE)}
}
\value{
matrix containing source sample, target sample, relation, contamination level, flip_flag
}
\description{
This function calculates contamination level, ratio of VAF_target to VAF_source
}
\references{
{TBA}
}
