% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/multipleSource.R
\name{multipleSource}
\alias{multipleSource}
\title{Identify true source from multisource contamination}
\usage{
multipleSource(sample_pairs,VAFdata,VAFcov,VAF_cutoff,VAF_cutoff1,p.val_cutoff,output_path)
=======
% Please edit documentation in R/multipleSource_noplot.R
\name{multipleSource}
\alias{multipleSource}
\title{Identify true source from multi source contamination}
\usage{
multipleSource(sample_pairs,VAFdata,VAF_cutoff,VAF_cutoff1,p.val_cutoff,output_path)
>>>>>>> 3b82f69fbb478a4e229c61fd993f97a6b55130cd
}
\arguments{
\item{sample_pairs}{data frame, variant allele frequency (VAF)}

\item{VAFdata}{data frame, VAF}

<<<<<<< HEAD
\item{VAFcov}{data frame, variant coverage}

=======
>>>>>>> 3b82f69fbb478a4e229c61fd993f97a6b55130cd
\item{VAF_cutoff}{numeric, minimum VAF (default:0.002)}

\item{VAF_cutoff1}{numeric, minimum VAF to be considered a source mutation (default:0.05)}

\item{p.val_cutoff}{numeric, pval cutoff to be considered significant contamination source (default:0.05)}

\item{output_path}{character, output directory}
<<<<<<< HEAD
=======

\item{plotVAF_flag}{logical, TRUE: generate contamination pairwise sample VAF scatter plot (default: FALSE)}
>>>>>>> 3b82f69fbb478a4e229c61fd993f97a6b55130cd
}
\value{
list, remaining paiwise sample relationship information after Fisher exact test that eliminates insignificant contamination pair
}
\description{
This function uses Fisher exact test to filter multiple sources that contaminate a particular target sample
}
\references{
{TBA}
}
