% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairCommon.R
\name{pairPCommon}
\alias{pairPCommon}
\title{Pairwise Sample Function}
\usage{
pairPCommon(VAFdata,VAF_cutoff,num_round_digit,n_sample)
pairShare(VAFdata,VAF_cutoff,n_sample)
pairList(VAFdata,n_sample,del)
pairCommList(SNPcomm,n_sample)
}
\arguments{
\item{VAFdata}{data frame, variant allele frequency (VAF)}

\item{VAF_cutoff}{numeric, minimum VAF}

\item{num_round_digit}{numeric, number of rounding decimal digit (default:3)}

\item{n_sample}{numeric, number of sample}

\item{SNPcomm}{matrix, output from the calling of pairPCommon(VAFdata,VAF_cutoff,num_round_digit,n_sample)}

\item{del}{character, file delimiter}
}
\value{
pairPCommon: matrix, nxn pcommon of pairwise sample, n: number of sample

pairShare: matrix, nxn number of common snps of pairwise sample

pairList: vector, all possible pairwise samples

pairCommList: list, all possible pairwise samples and the respective pcommon values
}
\description{
4 different functions for pairwise sample
}
\details{
pairPCommon calculates pcommon of pairwise sample

pairShare counts number of common SNPs for pairwise sample

pairList generates all pairwise sample list

pairCommList tabulates pairwise pcommon for each paired sample in a row
}
\references{
{TBA}
}
