% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circos_link.R
\name{plot_circos_link}
\alias{plot_circos_link}
\title{Plot circos with links to show contamination samples}
\usage{
plot_circos_link (file,R,W,plotsize,titleStr,seg.lab.size,fig.file, contaminatedOnly,sameSubjectOnly,allSamples,output_path)
}
\arguments{
\item{file}{character, circos data file name with its path}

\item{R}{integer, circle radius}

\item{W}{integer, segment width}

\item{plotsize}{integer, figure size}

\item{titleStr}{character, title of circos plot}

\item{seg.lab.size}{numeric, segment label font size}

\item{fig.file}{character, circos figure file name}

\item{contaminatedOnly}{logical, only plot contaminated pairs? (default: TRUE)}

\item{sameSubjectOnly}{logical, only plot same subject pairs? (default: FALSE)}

\item{allSamples}{character vector, sample ids}

\item{output_path}{character, directory}
}
\value{
.tif figure at output directory
}
\description{
This function reads in circos plotdata and creates circos plot
}
\references{
{TBA}
}
