% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionCountMutation.R
\name{regionCountMutation}
\alias{regionCountMutation}
\title{Count number of mutation in 7 regions of VAF scatter plot for a pair of samples (s1,s2)}
\usage{
regionCountMutation(sample_pairs,VAFdata,SNPcount,SNPshare,VAF_cutoff,VAF_ignore,n)
}
\arguments{
\item{sample_pairs}{data frame, pairwise samples information}

\item{VAFdata}{data frame, mutation VAF}

\item{SNPcount}{integer matrix, number of SNPs for each sample}

\item{SNPshare}{integer matrix, number of common SNPs}

\item{VAF_cutoff}{numeric, minimum VAF (default: 0.002)}

\item{VAF_ignore}{numeric, ignore variant below this cutoff (default: 0.2)}

\item{n}{integer, number of sample}
}
\value{
matrix, (n mutation x 10) containing number of SNPs in 7 regions, number of SNPs in s1, number of SNPs in s2, common SNPs between s1 and s2
}
\description{
This function counts number of SNPs in 7 regions of pairwise samples VAF scatter plot
}
\references{
{TBA}
}
