% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampleContamination.R
\name{run_sampleContamination}
\alias{run_sampleContamination}
\title{Main function to run sample contamination analysis}
\usage{
run_sampleContamination(data_path, output_path, config.file="config.txt",rmcov_cutoff=0,manualsetPara=FALSE)
}
\arguments{
\item{data_path}{character, VAF and COV data directory}

\item{output_path}{character, output directory}

\item{config_file}{character, configuration file}

\item{manualsetPara}{logical, TRUE: parameter setting base on distribution of pairwise sample commonality, FALSE: default setting as in configuration file (default:FALSE)}

\item{manual_localPcomm}{numeric, localPcomm for classification low/high contamination (default: 10)}

\item{manual_center}{numeric, manual center cutoff for same subject determination (default: 50)}

\item{rmcov_cutoff}{numeric, exclude SNPs with coefficient of variance (COV) below this cutoff}

\item{rmcov_cutoff}{numeric, TRUE: filter SNPs with low covariance of coefficient (COV), FALSE: (default:FALSE)}
}
\value{
contamination tab delimited text file with 3 columns: source sample, target sample, contamination percentage. A .tif circos plot figure
}
\description{
This function detects sample contamination base on variant allele frequency (VAF) and variant coverage
}
\references{
{TBA}
}
