% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParameterConfig.R
\name{setParameterConfig}
\alias{setParameterConfig}
\title{Set parameter with information from configuration file}
\usage{
setParameterConfig(filename)
}
\arguments{
\item{filename}{character, configuration file name and its path}
}
\description{
This function reads in configuration file to set parameters
}
\references{
{TBA}
}
