% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParameterConfig.R
\name{setParameterConfig}
\alias{setParameterConfig}
\title{Set parameters base on configuration file}
\usage{
setParameterConfig(configFile)
}
\arguments{
\item{configFile}{character, configuration file name and its path}
}
\description{
This function takes in data from configuration file to set parameters
}
\references{
{TBA}
}
